refactor: move EJS view templates to another folder
The default folder of EJS view templates is `views` related to the root
folder. Now we determine to use AMIS, a low-code rendering engine, and
we need to put the view configuration files here. So the EJS templates
are now stored in `views/ejs`, while the AMIS files in `views/amis`.