feat: add AMIS page for home
This commit adds an AMIS page for home page of this project. It only
contains a simple framework (a button toolbar, a user information board
and an announcement bulletin board) without useful information. It also
has some default settings from AMIS and most of them will be updated in
the future.

The visual editor of AMIS returns page configurations in pure JSON
format. But JavaScript format is more flexible and is better supported
on syntax highlighting. We use a simple website to convert JSON to JS.

These messages perhaps won't be included in future similar commits.

see: https://www.convertsimple.com/convert-json-to-javascript/