feat: add more utilities from hikarushare/tracker
We added more utilities from another repo "tracker" of our organization
for some specific uses. Now it's not more than a simple copy & paste.

We think the utilities may need to be moved to another new repo so that
other repos (like this one and "hikarushare/tracker") can import it via
NPM.